+++ app/plugins/router.js
@@ -1,6 +1,7 @@
 const routes = [].concat(
   require('../routes/home'),
   require('../routes/conversation'),
+  require('../routes/testing'),
   require('../routes/login'),
   require('../routes/reset'),
   require('../routes/healthy'),
