+++ app/routes/conversation.js
@@ -5,7 +5,6 @@ const { schemes } = require('../domain/schemes')
 const { getChatHistory, parseMessage } = require('../utils/langchain-utils')
 const { trackMessage, trackSystemMessage, trackConversationPageView } = require('../lib/events')
 const boom = require('@hapi/boom')
-const { logger } = require('../lib/logger')
 const { redact } = require('../utils/redact-utils')
 const config = require('../config')
 
@@ -136,71 +135,6 @@ module.exports = [
         conversationId,
         showHintText: true,
         schemesList
-      })
-    }
-  },
-  {
-    method: 'GET',
-    path: '/test_prompts',
-    handler: async (request, h) => {
-      if (config.endpointTestingEnabled !== true) {
-        return h.response().code(404)
-      }
-
-      logger.info('Testing prompts')
-
-      const inputs = [
-        'Give me grants for deer fencing',
-        'Give me grants to increase biodiversity',
-        'Give me grants for planting flowers',
-        "What's the price for FG10",
-        'what land types does AB1 apply to',
-        'funding for slurry'
-      ]
-
-      const processInput = async (input, index) => {
-        try {
-          const startTime = new Date()
-          let response
-          let passedValidation = false
-
-          try {
-            response = await fetchAnswer(request, input, [])
-            passedValidation = parseMessage(request, response) !== undefined
-          } catch (error) {
-            logger.info(`Response ${index + 1} out of ${inputs.length} failed validation or fetch with error:`, error)
-          }
-
-          const endTime = new Date()
-          const responseDuration = ((endTime.getTime() - startTime.getTime()) / 1000).toFixed(2)
-          const responseLength = response && passedValidation ? JSON.parse(response).answer.split(' ').length : response.split(' ').length
-
-          logger.info(`Response ${index + 1} out of ${inputs.length} generated in ${responseDuration} seconds`)
-
-          return {
-            question: input,
-            answer: passedValidation ? JSON.parse(response) : response,
-            responseDuration,
-            responseLength,
-            passedValidation
-          }
-        } catch (error) {
-          return {
-            question: input,
-            answer: 'Error:' + error,
-            responseDuration: 0,
-            responseLength: 0,
-            passedValidation: false
-          }
-        }
-      }
-
-      const responses = await Promise.all(inputs.map(async (input, index) =>
-        await processInput(input, index)
-      ))
-
-      return h.view('test_prompts_response', {
-        responses
       })
     }
   }
